backend/requirements.txt
bcrypt==4.0.0
black==22.8.0
blinker==1.5
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pymongo==4.2.0
python-dotenv==0.21.0
tomli==2.0.1
Werkzeug==2.2.2
WTForms==3.0.1
